requirements.txt
pytest==6.2.5
pytest-cov==3.0.0
Faker==9.7.1
mock==4.0.3
pyjwt==2.3.0
flask==3.0.0
gunicorn==20.1.0
flask-cors==3.0.10
celery==5.3.6
boto3==1.19.6
requests==2.26.0
bcrypt==3.2.0
pymongo==3.12.1
python-dotenv==1.0.0
flask-swagger-ui
